requirements.txt
pycryptodome>=3.19.0
flask>=2.3.0

